requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
langchain
langchain-groq
langchain-mongodb
langchain-chroma
langchain-huggingface
langchain-text-splitters
pymongo
python-dotenv
sentence-transformers
langchain-community
datasets
python-jose[cryptography]
passlib[bcrypt]
